Update VRC_Locator.PS1
Added check for ExecutionPolicy. Policy needs to be set to RemoteSigned for script to work. Also adding batch files to repo to start the ps1 file as admin and with out admin. Admin will be needed to set the ExecutionPolicy if it immediately closes when running it for the first time.